Revert "Set RELEASE_AIDL_USE_UNFROZEN false for 'next'"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1dc98ebe-58ad-4297-92fe-32c2a70070f0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I81980218ee2ee11307203f03c5a230096b745260