Revert "Add RELEASE_EXPOSE_FLAGGED_API release config flag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e219a77c-66ba-4c4b-9652-f5edfb56a726). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I45b498bcb4881570534c7135d522fb6118d2703c