Revert "Add RELEASE_EXPOSE_FLAGGED_API release config flag"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e093df71-66ae-4eb4-baa0-754c89128ba5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6b29ec0c279a747fddaf7e1bb0ac08d6a6c2ed7f