Revert "Revert^2 "Add RELEASE_EXPOSE_FLAGGED_API release config ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8c999e58-33dd-4e89-9587-bd5581f97ead). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ice62952df3b94aa2aa64c348b9775442f9fc0ee7