Revert "Revert^2 "Add RELEASE_EXPOSE_FLAGGED_API release config ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b906ac9e-e163-4549-ae55-92532184e81b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I124e78389155ffe8274bd544f2666f8f96f6b24b