Enable AVF related build time flags in trunk_staging
Features these flags are guarding are in stable state to be enabled in
trunk_staging. We do not intend to enable them further any time soon, we
just want them to be covered by the CI tests in presubmit :)

So far we are not enabling the RELEASE_AVF_ENABLE_LLPVM_CHANGES, we will
enable it separately later once we verify that the feature is ready to
enabled in trunk_staging.

Bug: 301951842
Test: atest --test-mapping packages/modules/Virtualization:avf-presubmit
Change-Id: Ic4b4d315b14d8bea028aa95533bc5b2c6059f3eb