Revert "Set using mainline prebuilts (default) for next target only."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8b21a0c8-774c-4b62-aa0e-47dce90c5f98). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3850dd63374df72f16c431e1da504df69acc303d