Revert "Set using mainline prebuilts (default) for next target only."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1ab792c3-75b7-489b-aae0-5f63ceb32c34). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia8e4d94d5d3ba9e759d44569c2b19681c9eda47e