Revert "Set using mainline prebuilts (default) for next target only."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cd740c1b-43a2-4170-80b6-49c56c3605fe). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I10edd59cf2109309ba4bf9ac26483901276c6162