Revert "Set using mainline prebuilts (default) for next target only."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ae1187fd-8215-46d0-949c-7099878d54f8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I853c8e4121d8ae4ef164557c63c0b5bd2dcc03a1