Revert "Set using mainline prebuilts (default) for next target only."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f98acf34-2f11-4003-8b3f-2a29edda1839). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idc6227f2ca8568e37627af8d8b4409fb2f0524ed