Revert "Set using mainline prebuilts (default) for next target only."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1dd6daa8-0a62-4746-803a-8fb058dc9939). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icb503a566a5303ef750bd510868213f3fcf4303f