Revert "Set using mainline prebuilts (default) for next target only."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/881b82d2-ae28-497d-bad9-0ebeac956f97). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifd1abab14a408610296c03cdf32beed2a8fd67d2