Revert "Set using mainline prebuilts (default) for next target only."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4ad03a9-46c8-44ac-8f7a-5520a7a4cd10). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaac279b97f4eed0a5eba20046e20805a8355f6d7