Revert "Set using mainline prebuilts (default) for next target only."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/18e2205c-64b4-4244-bd6a-f5604cf9e70b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia967631352b64f9408ee94cdfc90628b01ca1f07