Revert "Set using mainline prebuilts (default) for next target only."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a195fc65-8e7f-45ad-81cf-382764b34293). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5db641159218b8a5d5bee2bda76296304c9ee0b0