Chagne release configuration Android.bp structure
To support local override, add variable ACONFIG_VALUES_LIST_LOCAL in the
top level Android.bp file, so it will be visible to the Android.bp file
in the subdirectory.

The variable ACONFIG_VALUES_LIST_LOCAL will be used in the subdirectory
as part of the aconfig value set.

If the variable ACONFIG_VALUES_LIST_LOCAL is set the release
configuratin will take the values from it. If it is empty, it will only
take values from their ACONFIG_VALUES_LIST.

Bug: 298692416
Test: source build/envsetup.sh && overrideflags, and then check the
output

Change-Id: Ie54e46adf0019bf153a93dede2fb1dc14f0c4d23