Revert "Deprecate VNDK from all trunk builds"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ccaf1743-d1a1-42a5-b41f-e4e287d654f4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie9cc28282a4f5ec19bcb9fbe2e22e3122cc6116f