Add VanillaIceCream to active codename list
The RELEASE_PLATFORM_VERSION_ALL_CODENAMES is meant to describe all
preview codenames and is reference in build/soong to compute the valid
preview list.

This is required to support handling modules that have a min_sdk_version
of VIC (for instance, com.android.profiling) that would otherwise fail
processing on -next release configurations as the preview codename would
be considered invalid.

Bug: 316638747
Test: lunch aosp_arm64-next-userdebug && m com.android.profiling
Change-Id: I71ceb155d44fa8eb07ab4e017133667b13dc777d