Introduce a root release config
Add a new release config, root, that all other release configs inherit
from.

This release config is used to set the values for the
com.android.aconfig.test flags, which are used in aconfig integration
tests.

Other changes to the root release config are not expected: you probably
want to modify one of the next or trunk release configs intead.

Test: printflags --filter=fully_qualified_name:com.android.aconfig.test.enabled_ro # manually verify that the flag is READ_ONLY + ENABLED
Bug: 317178122
Change-Id: I375d53e30f856829a6e0eab7623a168e402b7d49