Revert "Stage the new flag emoji font"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/22c373dd-a687-4f3e-94e0-be20655136ee). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9bfb3c0461cff12a98fd6fdc1137e458f5506024