Revert "Enable flag: enable avf virt cpufreq in trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1ccd2fc3-f8b5-4328-bddc-a92e425ba59c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If3aaf4d1bbecacae16fe5e49244cc248d8460624