Revert "Enable flag: enable avf virt cpufreq in trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dd358b91-86c3-43ae-88a3-540b2118b814). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib54d56d9c757d96fac75d390ebd3b243c9cf6bf8