Revert "Enable flag: enable avf virt cpufreq in trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/49371e17-9923-49da-83f4-7f9d3d7992eb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I153c3d960e79c63fb3f5755f2f860d5cd43c4ee3