Revert "Add microfuchsia flag declaration and value"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e1c0223f-4063-4c41-bafe-014da5351387). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 361108146
Change-Id: I2ef3e50bfcae4d47d5e8cd34d0b77d161c768f25
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com