BPF: Define build time feature flag for libbpf
For trunk stable development, we use a build time feature flag for the
libbpf functionality to be added to bpfloader.

This approach will allow merging and testing of libbpf functionality
without changing the binaries produced with the feature disabled. Build
time flagging is required because bpfloader runs early in the init
process, and runtime flags likely aren't enabled that early in boot.

Bug: 359646531
Test: manual/treehugger
Change-Id: Ib8cf81029ddcf783c2a8e6824dfdc6393a4a80dd
Signed-off-by: Neill Kapron <nkapron@google.com>