Revert "Add new AVF flags & enable them in trunk-staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/547300d0-f6e4-4047-82f5-9f83913cf1c4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifcd4c08ac965304271a0cee5dfc72afd9463173b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com