Revert "Enable RELEASE_USE_TRANSITIVE_JARS_IN_CLASSPATH flag in ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7fcc0aeb-e4cc-48eb-b6a1-b083b2a71227). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 365146337
Change-Id: I8c7590eb033b9c2fe369413d95a1bbe85a7913c2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com