Revert "Enable RELEASE_USE_TRANSITIVE_JARS_IN_CLASSPATH flag in ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e4cf01b0-9fbc-4dd6-aa94-14a88f353cba). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 365146337
Change-Id: Ic8685eb0398fb500be14095716fcbdf432b31ce1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com