Revert "Allow trunk_staging approval for release managers"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c0fc806d-58ba-4cf9-b109-57a62cfbab3e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I394f230a3bcd1ea8c5b8e3b97c39bd73935a1cd2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com