Add build flag for fingerprint.
This flag will guard fingerprinting package.map files for aconfig flags
so that we can tell if we are reading from a file built at a different
time than the read libs were built (in the case of cross-container
read). We need this to be a build-time flag as we are guarding the
aconfig code that writes the storage files at build time (as well as
some codegen logic).

Bug: 316357686
Test: n/a, this change just adds the flag declaration
Change-Id: Ib0af421f60ed5ecc3e744a184cccd24448e1f937