Revert "Revert^2 "Enable RELEASE_USE_OPTIMIZED_RESOURCE_SHRINKIN..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c34184b6-c2b7-46e7-9f3e-8bc6f406b3ba). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 370691134
Change-Id: Icd70bdcc0eb482c711ffe520120e90efec68ac98
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com