Revert "Enable RELEASE_HIDDEN_API_EXPORTABLE_STUBS in trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/178ef64c-6257-422b-b1d2-4cfba5dc3d35). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 371226454
Change-Id: I450796b78d335d8ec50df66eea763e135025ddae
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com