Revert "Enable RELEASE_HIDDEN_API_EXPORTABLE_STUBS in trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fa6a133f-ea0a-4af3-8fae-ba34fda9e403). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 371226922
Bug: 371229475
Bug: 371231272
Bug: 371232400
Change-Id: I8a3efb99fd7089534755255cbe0aec27d4faf5f8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com