Revert "Enable RELEASE_HIDDEN_API_EXPORTABLE_STUBS in trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1d57ec10-c526-4d96-bfa9-f476c43e4812). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 371391168
Change-Id: I58a2ab49bcd68d6ae54c612d69a29bb47b5df8cd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com