Revert "Enable RELEASE_HIDDEN_API_EXPORTABLE_STUBS in trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/41d34786-09bf-4ff4-bac2-c7fe43cde295). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 371395847
Change-Id: I64dfdc82685de5f7cdaf31e2591250033c662ab8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com