Revert "Enable RELEASE_HIDDEN_API_EXPORTABLE_STUBS in trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/30082e65-3493-4003-b261-93af932dd656). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 371391382
Change-Id: I485227a1b09684a052dbbd73466475f0d0abfd34
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com