Revert "Enable RELEASE_HIDDEN_API_EXPORTABLE_STUBS in trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f0311ba8-316f-4416-aad5-391b3bc22043). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 371244020
Change-Id: I1af05ebca32b5fa78e58565ab39c7f4bd19efa81
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com