Revert "Enable RELEASE_HIDDEN_API_EXPORTABLE_STUBS in trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/340c6ed3-ac62-4168-b34d-547885fc797b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 371409586
Change-Id: Id26d0722860f06eacaea2a216f2f0d4c5e679162
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com