Revert "Enable RELEASE_HIDDEN_API_EXPORTABLE_STUBS in trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1e967b3c-2341-4c43-9f2b-38a97eb6f7ed). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 371366779
Change-Id: Iba06afc1a3b1860c91fafd090b95fbf02c8bcb4e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com