Revert "Enable RELEASE_HIDDEN_API_EXPORTABLE_STUBS in trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/224c5b9a-ec9d-4777-aeb8-2482457f422d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 371427504
Bug: 371429715
Bug: 371446965
Bug: 371449803
Change-Id: I345999f50b0b6bfeb6823b8dae493c2403c0ebae
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com