Revert "Move ap3a configuration to build/release."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/af4e4db5-a835-495e-a4fa-5cb80190991f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 371479194
Change-Id: I096a16c70a936e52f1c57e4b73d4273842eeab00
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com