Revert "Revert^3 "Enable RELEASE_USE_OPTIMIZED_RESOURCE_SHRINKIN..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/791903a8-f5b1-4fec-94e4-5bdbe0fcc679). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibe825fd3b84c73c1977869a351d10cb60ceb22c1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com