Revert "Add RELEASE_NFC_MAINLINE_MODULE flag to trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d0aa484a-55c7-4ab2-bfe0-5f39af0bde72). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 373899734
Change-Id: I5ab2fb7db64b9d779d4b7a29bf5b86f8d4040c04
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com