Revert "Add RELEASE_NFC_MAINLINE_MODULE flag to trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/55409236-33fa-4121-95ae-3efe22fce9a0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 373926404
Change-Id: I28633ea8a85648f4664644d8806a53f0f188c7e2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com