Revert "Add RELEASE_NFC_MAINLINE_MODULE flag to trunk_staging"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d22461a1-bb6b-4e90-a258-8b0def81b589). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 373949628
Change-Id: Ia43c0ba1acf5de2df74db242c8e6004d1a176db3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com