Revert "Revert^5 "Enable RELEASE_USE_OPTIMIZED_RESOURCE_SHRINKIN..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3ea9a8c0-e196-42fb-99ae-6ff684a5b895). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iecb2771181dee65ec0f224c5bf796a2784254bdc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com