Revert "Revert^5 "Enable RELEASE_USE_OPTIMIZED_RESOURCE_SHRINKIN..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4f3c8361-ddd3-4951-bcb0-e1d8c984f5d0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If0991f2142381b1b6128a18f0cd07c4959a1ae26
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com