Revert "Update Emoji to 4.074 which has latest Unicode 16 emojis"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/02f0f361-6663-45b6-836b-12973eb7cb0e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 374686442
Change-Id: Iece262c4b8d7270e81850d6784cbacdae894f294
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com