Revert "Revert "Update Emoji to 4.074 which has latest Unicode 1..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8699f2f8-baa0-4dbc-8b31-dca1b457b13a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I88d55fc9507210bd005bd5ff0f1533e9a2db6931
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com