Revert "Move ap3a configuration to build/release."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4d283a8-c7f7-4d51-ba80-44c049f7773f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia1c7aa51478e01e23e5d9dfb6bc2f91bcc06e15f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com