Revert "trunk_staging: set codename to Baklava"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a6f53a4b-2a6a-42a7-bab5-e018daf1a823). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I83b29eb957d5eef7fdffbefb393718c6a591da15
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com