Revert "Add flag to control if to target java 21 by default"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f4f2d269-8064-4bf2-a427-16bc034f4055). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia4f85131458ff83453b231cc951e4725352246f0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com