Revert "Move P24 and CT3 Kernel flag declaration here"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/66cc9b59-3ecf-42d9-8c9b-21cfae803073). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6d30e76372aedc20c11fc143792548642496ba85
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com